+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Implement a new shortcut key assignable command: "Visit the canonical URL of the current tab if available".
 - Rename the feature "Share URL" to "Save URL online".
 - Add a shortcut key "C" for copying the canonical URL in the popup.
+- Copy a Markdown link to the current tab page when the context menu item "Copy a Markdown link to the current page" is invoked from within the popup.
 
 ## 0.9.2